LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/urls/resolving-urls/query-encoding/windows-1251-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/echo-query?%C3%A5' failed: WebSocket is closed before the connection is established.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: Did not parse stylesheet at 'http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/stash.py?q=%C3%A5&action=put&id=8ff7b824-ce33-468b-b199-6b3b918a2c88' because non CSS MIME types are not allowed in strict mode.
CONSOLE MESSAGE: Did not parse stylesheet at 'http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/stash.py?q=%C3%A5&action=put&id=f7bdab7a-cd42-439c-9f9a-0b1aa69e9c93' because non CSS MIME types are not allowed in strict mode.
CONSOLE MESSAGE: Did not parse stylesheet at 'http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/stash.py?q=%C3%A5&action=put&id=b1220246-6f7d-42dd-b25c-9618e11613be' because non CSS MIME types are not allowed in strict mode.

Harness Error (TIMEOUT), message = null

PASS load nested browsing context <frame src>
PASS load nested browsing context <iframe src>
TIMEOUT load nested browsing context <object data> Test timed out
TIMEOUT load nested browsing context <embed src> Test timed out
PASS loading css <link>
PASS loading js <script>
FAIL loading image <img src> assert_equals: expected substring %26%23229%3B got unknown query expected (undefined) undefined but got (number) 256
TIMEOUT loading image <embed src> Test timed out
TIMEOUT loading image <object data> Test timed out
FAIL loading image <input src> assert_equals: expected substring %26%23229%3B got unknown query expected (undefined) undefined but got (number) 256
FAIL loading image <video poster> assert_equals: expected substring %26%23229%3B got unknown query expected (undefined) undefined but got (number) 256
FAIL loading video <video> assert_equals: expected substring %26%23229%3B got unknown query expected (undefined) undefined but got (number) 300
FAIL loading video <video><source> assert_equals: expected substring %26%23229%3B got unknown query expected (undefined) undefined but got (number) 300
FAIL loading video <audio> assert_equals: expected substring %26%23229%3B got unknown query expected (undefined) undefined but got (number) 300
FAIL loading video <audio><source> assert_equals: expected substring %26%23229%3B got unknown query expected (undefined) undefined but got (number) 300
PASS loading webvtt <track>
PASS submit form <form action>
PASS submit form <input formaction>
PASS submit form <button formaction>
FAIL <base href> assert_true: expected substring %26%23229%3B got http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/resource.py?q=%C3%A5&encoding=windows-1251&type= expected true got false
PASS Worker constructor
FAIL SharedWorker constructor Can't find variable: SharedWorker
PASS EventSource constructor
PASS EventSource#url
PASS window.open()
FAIL <a>.search assert_true: href content attribute expected substring %26%23229%3B got http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/resource.py?q=%C3%A5&encoding=windows-1251&type=html expected true got false
FAIL <area>.search assert_true: href content attribute expected substring %26%23229%3B got http://localhost:8800/html/infrastructure/urls/resolving-urls/query-encoding/resources/resource.py?q=%C3%A5&encoding=windows-1251&type=html expected true got false
FAIL history.pushState assert_equals: url was resolved against the iframe's URL instead of the settings object's API base URL expected -1 but got 76
FAIL history.replaceState assert_equals: url was resolved against the iframe's URL instead of the settings object's API base URL expected -1 but got 76
PASS SVG <a>
TIMEOUT SVG <feImage> Test timed out
FAIL SVG <image> assert_equals: expected "%26%23229%3B" but got "%C3%A5"
PASS SVG <use>
FAIL XMLHttpRequest#open() assert_equals: expected "%C3%A5" but got "%26%23229%3B"
PASS importScripts() in a dedicated worker
FAIL Worker() in a dedicated worker assert_equals: expected "%C3%A5" but got "ReferenceError: Can't find variable: Worker"
FAIL SharedWorker() in a dedicated worker assert_equals: expected "%C3%A5" but got "ReferenceError: Can't find variable: SharedWorker"
FAIL importScripts() in a shared worker Can't find variable: SharedWorker
FAIL Worker() in a shared worker Can't find variable: SharedWorker
FAIL SharedWorker() in a shared worker Can't find variable: SharedWorker
PASS WebSocket constructor
PASS WebSocket#url
PASS Parsing cache manifest (CACHE)
PASS Parsing cache manifest (FALLBACK)
PASS Parsing cache manifest (NETWORK)
FAIL CSS <link> (windows-1251) #<id> { background-image:<url> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
PASS CSS <link> (utf-8) #<id> { background-image:<url> }
FAIL CSS <style> #<id> { background-image:<url> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
FAIL CSS <link> (windows-1251) #<id> { border-image-source:<url> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
PASS CSS <link> (utf-8) #<id> { border-image-source:<url> }
FAIL CSS <style> #<id> { border-image-source:<url> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
FAIL CSS <link> (windows-1251) #<id>::before { content:<url> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
PASS CSS <link> (utf-8) #<id>::before { content:<url> }
FAIL CSS <style> #<id>::before { content:<url> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
FAIL CSS <link> (windows-1251) @font-face { font-family:<id>; src:<url> } #<id> { font-family:<id> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
PASS CSS <link> (utf-8) @font-face { font-family:<id>; src:<url> } #<id> { font-family:<id> }
FAIL CSS <style> @font-face { font-family:<id>; src:<url> } #<id> { font-family:<id> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
FAIL CSS <link> (windows-1251) #<id> { display:list-item; list-style-image:<url> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
PASS CSS <link> (utf-8) #<id> { display:list-item; list-style-image:<url> }
FAIL CSS <style> #<id> { display:list-item; list-style-image:<url> } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
PASS CSS <link> (windows-1251) @import <url>;
PASS CSS <link> (utf-8) @import <url>;
PASS CSS <style> @import <url>;
FAIL CSS <link> (windows-1251) #<id> { cursor:<url>, pointer } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
PASS CSS <link> (utf-8) #<id> { cursor:<url>, pointer }
FAIL CSS <style> #<id> { cursor:<url>, pointer } assert_equals: expected "%C3%A5" but got "%26%23229%3B"
FAIL <?xml-stylesheet?> (CSS) assert_equals: expected "\"%C3%A5\"" but got "\"%26%23229%3B\""
PASS URL constructor, url
PASS URL constructor, base
PASS Scheme ftp (getting <a>.href)
PASS Scheme file (getting <a>.href)
FAIL Scheme gopher (getting <a>.href) assert_true: expected substring %26%23229%3B got gopher://example.invalid/?x=%C3%A5 expected true got false
PASS Scheme http (getting <a>.href)
PASS Scheme https (getting <a>.href)
PASS Scheme ws (getting <a>.href)
PASS Scheme wss (getting <a>.href)
PASS Scheme mailto (getting <a>.href)
PASS Scheme data (getting <a>.href)
PASS Scheme javascript (getting <a>.href)
PASS Scheme ftps (getting <a>.href)
PASS Scheme httpbogus (getting <a>.href)
PASS Scheme bitcoin (getting <a>.href)
PASS Scheme geo (getting <a>.href)
PASS Scheme im (getting <a>.href)
PASS Scheme irc (getting <a>.href)
PASS Scheme ircs (getting <a>.href)
PASS Scheme magnet (getting <a>.href)
PASS Scheme mms (getting <a>.href)
PASS Scheme news (getting <a>.href)
PASS Scheme nntp (getting <a>.href)
PASS Scheme sip (getting <a>.href)
PASS Scheme sms (getting <a>.href)
PASS Scheme smsto (getting <a>.href)
PASS Scheme ssh (getting <a>.href)
PASS Scheme tel (getting <a>.href)
PASS Scheme urn (getting <a>.href)
PASS Scheme webcal (getting <a>.href)
PASS Scheme wtai (getting <a>.href)
PASS Scheme xmpp (getting <a>.href)
PASS Scheme web+http (getting <a>.href)
